requirements.txt
fake_useragent==2.2.0
gspread==6.2.1
langchain_core==0.3.75
langchain_google_genai==2.1.10
oauth2client==4.1.3
pandas==2.3.2
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
selenium==4.35.0
undetected_chromedriver==3.5.5
webdriver_manager==4.0.2
